Replaced Vmt parser with KeyValue lib
Replaced (bad) vmt parser with KeyValue library
Pre-allocated space in resource manager
Removed some old comments
